--- a/airflow-core/src/airflow/api_fastapi/core_api/routes/ui/grid.py
+++ b/airflow-core/src/airflow/api_fastapi/core_api/routes/ui/grid.py
@@ -163,12 +163,12 @@ def get_dag_structure(
     serdags = session.scalars(
         select(SerializedDagModel).where(
             SerializedDagModel.dag_id == dag_id,
+            SerializedDagModel.id != latest_serdag.id,
             SerializedDagModel.dag_version_id.in_(
                 select(TaskInstance.dag_version_id)
                 .join(TaskInstance.dag_run)
                 .where(
                     DagRun.id.in_(run_ids),
-                    SerializedDagModel.id != latest_serdag.id,
                 )
             ),
         )